requirements.txt
# PyPI dependencies
accept-types==0.4.1
aiohttp==3.8.3
aiokafka==0.7.2
dataclasses-avroschema==0.41.1
isodate==0.6.1
jsonschema==3.2.0
networkx==2.5
protobuf==3.19.5
python-dateutil==2.8.2
tenacity==8.0.1
typing_extensions==4.5.0
ujson==5.7.0
watchfiles==0.19.0

# Patched dependencies
#
# Note: It is important to use commits to reference patched dependencies. This
# has two advantages:
# - Reproducible builds
# - The contents of the file change, which invalidates the existing docker
#   images and forces a new image generation.
#
git+https://github.com/aiven/kafka-python.git@1b95333c9628152066fb8b1092de9da0433401fd

git+https://github.com/aiven/avro.git@f37a4ea69560d798cc718fccd7f75223527f5ebb#subdirectory=lang/py


# Indirect dependencies
aiosignal==1.2.0 # aiohttp
async-timeout==4.0.2 # aiohttp
attrs==21.4.0 # jsonschema
certifi==2022.12.7 # requests, urllib3
chardet==3.0.4 # requests
charset-normalizer==2.0.11 # aiohttp, requests
decorator==5.1.1 # networkx
frozenlist==1.3.1 # aiohttp, aiosignal
idna==3.3 # aiohttp, requests, urllib3
lz4==3.0.2 # kafka
multidict==6.0.2 # aiohttp, yarl
python-snappy==0.6.1 # kafka
pyrsistent==0.18.1 # jsonschema
requests==2.27.1 # jsonschema
six==1.16.0 # dateutil
urllib3==1.26.8 # requests
yarl==1.8.1 # aiohttp
zstandard==0.18.0 # kafka
